better packet handling, no forced connect order
have mwconnector check the packet being received to make sure it is
intended for it. this consists of a few things:
- the packet is coming from the correct socket
- the packet is a string
- the packet begins with a number of newlines followed by an opening
curly brakcet
i can't just read a string and then check if that's valid, because GM
will throw a fit if it reads past the end of the buffer

all this work means that i was able to allow connecting to multitroid
before multiworld. despite how little code changed, this was a headache
due to GM not generating disconnect events on a disconnect, but still
re-using socket IDs